+++ Apps/Examples/Examples/SwiftUI Examples/ViewAnnotationsExample.swift
@@ -29,6 +29,7 @@ struct ViewAnnotationsExample: View {
             }
             .allowOverlap(allowOverlap)
             .selected(selected)
+
             // Dynamic view annotations, appeared on tap.
             // The anchor can point to bottom, top, left, or right direction.
             ForEvery(taps) { tap in
@@ -77,7 +78,6 @@ struct ViewAnnotationsExample: View {
             .onAnchorChanged { self.etaAnnotationAnchor = $0.anchor }
             .selected(true)
         }
-        .presentsWithTransaction(true) // Synchronize Metal and CALayer for better VA performance.
         .onMapTapGesture { context in
             taps.append(Tap(coordinate: context.coordinate))
         }